nghttp: attach transport to request before trying an actual connect
Previously, if an error occured early enough, e.g. during a socket connect call or during an ssl handshake, the associated transport object wouldn't be attached to the requesting session.
When the request manager saw an error, it should have been able to close the transport, release it to the connection pool, and try again.
Instead, the connection was never released, and depending on the config chosen, any retries would be starved because no connections had been returned to the pool.
This ensures that a transport object is attached, then a connect is attempted.
If an error occurs, the transport object will be (possibly) reset and released back to the pool, and retries will continue firing as applicable.